+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0
+
+* BREAKING CHANGES!
+* Support customize each rating widget.
+* Support builder constructor.
+
 ## 1.0.0
 
 * Release